[DataInfra][xplat][ldText][1/n]Create the implementation of ldText
This CR:

adds the implementation of ldText, which is the internal version of strings used in texts for public display. This ldText should be used in all times, in replacement of traditional, built-in string in Swift. 

creates method to call ldGetText() to get the string (used the same as traditional vanilla string). For example, ldGetText("ABC") returns "ABC". 

creates method to call ldGetTextDescription(), which can return the description of the text. For example, ldGetTextDescription("Time for lunch!") can return "This is an alert that remind user it is lunchtime". The purpose of this is to document the purpose and context of this text, paving the way for translations in the future. 

Text plan:

To be implemented. Currently, no code is calling this class, so it should not impact any existing features.

Privacy Review:

Since this class is for LD internal use, we should keep this confidential. We should not store any customer-sensitive data into this class, so that translators will not see confidential information.